+++ CHANGELOG.md
@@ -3,6 +3,13 @@
 We'll note all notable changes in this file, including bug fixes, enhancements, and all closed issues.
 Dates are in `YYYY-MM-DD` format and versions are in [semantic versioning](http://semver.org/) format.
 
+## 0.9.34 2024-02-24
+
+## Added
+
+-   [#343](https://github.com/wordplaydev/wordplay/issues/343): Improved login feedback.
+-   [#37](https://github.com/wordplaydev/wordplay/issues/37): Added match, e.g., `number ??? 1: 'one' 2: 'two' 'other'
+
 ## 0.9.33 2024-02-19
 
 ## Added